requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
transformers
torch
requests
accelerate
Pillow
